dashboard/mock/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
bcrypt==4.1.2 